+++ src/block-kit/Actions.tsx
@@ -4,12 +4,26 @@ import { JSXSlack } from '../jsx'
 import { ArrayOutput, ObjectOutput } from '../utils'
 import { BlockComponentProps } from './Blocks'
 import { ButtonProps } from './interactive/Button'
-import { SelectPropsBase } from './interactive/Select'
+import { SingleSelectPropsBase } from './interactive/Select'
 import { OverflowProps } from './interactive/Overflow'
+import { DatePickerProps } from './interactive/DatePicker'
 
 interface ActionsProps extends BlockComponentProps {
-  children: JSXSlack.Children<ButtonProps | SelectPropsBase | OverflowProps>
+  children: JSXSlack.Children<
+    ButtonProps | SingleSelectPropsBase | OverflowProps | DatePickerProps
+  >
 }
+
+const actionTypes = [
+  'button',
+  'static_select',
+  'external_select',
+  'users_select',
+  'conversations_select',
+  'channels_select',
+  'overflow',
+  'datepicker',
+]
 
 export const Actions: JSXSlack.FC<ActionsProps> = props => {
   const elements = JSXSlack(<ArrayOutput>{props.children}</ArrayOutput>)
@@ -18,6 +32,9 @@ export const Actions: JSXSlack.FC<ActionsProps> = props => {
     throw new Error(
       `The number of passed elements (${elements.length}) is over the limit. <Actions> block allows to include up to 25 elements.`
     )
+
+  if (elements.some(({ type }) => !actionTypes.includes(type)))
+    throw new Error(`<Actions> block has an incompatible element as children.`)
 
   return (
     <ObjectOutput<ActionsBlock>