+++ test/index.tsx
@@ -37,7 +37,20 @@ import JSXSlack, {
 beforeEach(() => JSXSlack.exactMode(false))
 
 describe('jsx-slack', () => {
-  describe('Block Kit as component', () => {
+  describe('Container components', () => {
+    describe('<Blocks>', () => {
+      it('throws error when <Blocks> has unexpected element', () =>
+        expect(() =>
+          JSXSlack(
+            <Blocks>
+              <b>unexpected</b>
+            </Blocks>
+          )
+        ).toThrow())
+    })
+  })
+
+  describe('Block Kit components', () => {
     describe('<Section>', () => {
       const section: SectionBlock = {
         type: 'section',
@@ -105,16 +118,23 @@ describe('jsx-slack', () => {
         }
       })
 
-      it('output section block with action accessories', () => {
+      it('outputs section block with action accessories', () => {
         for (const accessory of [
           <Button>Button</Button>,
           <Select>
-            <Option value="sel">Static select</Option>
+            <Option value="select">Static select</Option>
+          </Select>,
+          <Select multiple>
+            <Option value="select">Multiple select</Option>
           </Select>,
           <ExternalSelect />,
+          <ExternalSelect multiple />,
           <UsersSelect />,
+          <UsersSelect multiple />,
           <ConversationsSelect />,
+          <ConversationsSelect multiple />,
           <ChannelsSelect />,
+          <ChannelsSelect multiple />,
           <Overflow>
             <OverflowItem>Overflow</OverflowItem>
             <OverflowItem>item</OverflowItem>
@@ -124,10 +144,7 @@ describe('jsx-slack', () => {
           expect(
             JSXSlack(
               <Blocks>
-                <Section blockId="with_image">
-                  Accessory test
-                  {accessory}
-                </Section>
+                <Section blockId="with_image">test {accessory}</Section>
               </Blocks>
             )
           ).toStrictEqual([
@@ -135,6 +152,60 @@ describe('jsx-slack', () => {
               accessory: expect.objectContaining({ type: expect.any(String) }),
             }),
           ])
+        }
+      })
+
+      it('outputs section block with multi-select menus', () => {
+        // Static multiple select
+        const [s] = JSXSlack(
+          <Blocks>
+            <Section>
+              Select
+              <Select multiple maxSelectedItems={2} value={['a', 'c']}>
+                <Option value="a">a</Option>
+                <Option value="b">b</Option>
+                <Option value="c">c</Option>
+              </Select>
+            </Section>
+          </Blocks>
+        )
+
+        expect(s.accessory.type).toBe('multi_static_select')
+        expect(s.accessory.max_selected_items).toBe(2)
+        expect(s.accessory.initial_options).toHaveLength(2)
+
+        // Multiple select for external sources
+        for (const accessory of [
+          <ExternalSelect
+            multiple
+            maxSelectedItems={2}
+            initialOption={<Option value="a">a</Option>}
+          />,
+          <UsersSelect multiple maxSelectedItems={2} initialUser="U00000000" />,
+          <ConversationsSelect
+            multiple
+            maxSelectedItems={2}
+            initialConversation={['C00000000']}
+          />,
+          <ChannelsSelect
+            multiple
+            maxSelectedItems={2}
+            initialChannel="D00000000"
+          />,
+        ]) {
+          const [ms] = JSXSlack(
+            <Blocks>
+              <Section>Select {accessory}</Section>
+            </Blocks>
+          )
+
+          expect(ms.accessory.type.startsWith('multi_')).toBe(true)
+          expect(ms.accessory.max_selected_items).toBe(2)
+
+          const initialKey: any = Object.keys(ms.accessory).find(k =>
+            k.startsWith('initial_')
+          )
+          expect(ms.accessory[initialKey]).toHaveLength(1)
         }
       })
     })
@@ -292,8 +363,7 @@ describe('jsx-slack', () => {
       })
 
       it('outputs actions block with styled <Button>', () => {
-        // TODO: Remove type casting when supported style field on @slack/types
-        const buttonAction = (action as Function)(
+        const buttonAction = action(
           {
             type: 'button',
             text: { type: 'plain_text', text: 'Default', emoji: true },
@@ -723,6 +793,19 @@ describe('jsx-slack', () => {
           )
         ).toStrictEqual([buttonAction])
       })
+
+      it('throws error when passed selectable element with "multiple" prop', () =>
+        expect(() =>
+          JSXSlack(
+            <Blocks>
+              <Actions>
+                <Select multiple>
+                  <Option value="error">error</Option>
+                </Select>
+              </Actions>
+            </Blocks>
+          )
+        ).toThrow())
 
       it('throws error when the number of elements is 26', () =>
         expect(() =>